# swagger-spec.yml
info:
  title: "Node Express Server"
  version: "1.0.0"
  description: "Node Express Server With User Creation, Authentication, Password Reset using Email Verification and JWT's."
openapi: "3.0.0"
paths:
  "/user/users":
    get:
      description: "Gets all users."
      tags:
        - User
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              example:
                success: true
                message: "Successfully fetched all users."
                data: []
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
...
  "/user/username-available":
    post:
      description: "Checks if a username is assigned to a user."
      tags:
        - User
      requestBody:
        description: "The username to query with"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
            example:
              username: "username123"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  usernameAvailable:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Username is available."
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  usernameAvailable:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid request body."
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  usernameAvailable:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "An error occurred. Error: something went wrong."
...
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        journals:
          type: array
          items:
            $ref: "#/components/schemas/Journal"
        journalCategories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              selected:
                type: boolean
    Journal:
      type: object
      properties:
        title:
          type: string
          example: "Journal 1"
        entry:
          type: string
          example: "Journal 1"
        category:
          type: string
        date:
          type: string
        selected:
          type: boolean
...